# src/job_pricing/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Dynamic Job Pricing Engine API
  version: 1.0.0
  description: |
    Production-ready API for AI-powered compensation intelligence platform.

    ## Features
    - Job pricing with confidence scoring
    - Mercer job library integration
    - SSG skills extraction
    - Market data aggregation
    - Historical trends analysis

    ## Authentication
    All endpoints require JWT Bearer token authentication except /auth endpoints.

    ## Rate Limiting
    - Free Tier: 100 requests/hour
    - Standard Tier: 1000 requests/hour
    - Enterprise Tier: Unlimited

  contact:
    name: API Support
    email: api-support@company.com
  license:
    name: Proprietary
servers:
  - url: https://api.jobpricing.company.com/v1
    description: Production server
  - url: https://staging-api.jobpricing.company.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@company.com
                password:
                  type: string
                  format: password
                  example: SecurePassword123!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Generate new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'

  # ==================== Job Pricing ====================
...
  /users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      description: Retrieve authenticated user's profile information
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'

    put:
      tags:
        - Users
      summary: Update user profile
      description: Update authenticated user's profile information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                department:
                  type: string
                notification_preferences:
                  type: object
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            access_token:
              type: string
            refresh_token:
              type: string
            token_type:
              type: string
              example: Bearer
            expires_in:
              type: integer
              example: 3600
            user:
              $ref: '#/components/schemas/UserProfile'
    TokenResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            access_token:
              type: string
            token_type:
              type: string
              example: Bearer
            expires_in:
              type: integer
              example: 3600

    # Job Pricing Schemas
...
    UserProfile:
      type: object
      properties:
        user_id:
          type: string
        email:
          type: string
          format: email
        full_name:
          type: string
        department:
          type: string
        role:
          type: string
          enum: [Viewer, Analyst, Manager, Admin]
        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_INPUT
            message:
              type: string
              example: Validation error
            details:
              type: object
        metadata:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
              format: uuid